requirements.txt
pytz==2022.2
click==8.1.3
python-dotenv==0.21.0
requests==2.28.1